Rename component to targets.mutate
