+++ component/all/all.go
@@ -2,6 +2,6 @@
 package all
 
 import (
-	_ "github.com/grafana/agent/component/discovery/transformer" // Import discovery.transformer
-	_ "github.com/grafana/agent/component/local/file"            // Import local.file
+	_ "github.com/grafana/agent/component/local/file"     // Import local.file
+	_ "github.com/grafana/agent/component/targets/mutate" // Import targets.mutate
 )
