+++ component/targets/mutate/types.go
@@ -1,4 +1,4 @@
-package transformer
+package mutate
 
 import (
 	"fmt"
